% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annuity_present_value_factor.R
\name{annuity_present_value_factor}
\alias{annuity_present_value_factor}
\title{Present value factor of an annuity}
\usage{
annuity_present_value_factor(rate, periods)
}
\arguments{
\item{rate}{Interest rate.}

\item{periods}{Number of periods.}
}
\value{
Present value per dollar that is received in the future
}
\description{
Present value factor of an annuity
}
\examples{
present_value_factor(0.05, 2)
present_value_factor(0.10, 7)
}
