% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cagr.R
\name{cagr}
\alias{cagr}
\title{Calculate the compound annual/periodic growth rate}
\usage{
cagr(present_value, future_value, periods = 1)
}
\arguments{
\item{present_value}{Value of money at time t0.}

\item{future_value}{Value at maturity.}

\item{periods}{Number of periods.}
}
\value{
Compound annual/periodic growth rate (CAGR)
}
\description{
Calculate the compound annual/periodic growth rate
}
\examples{
cagr(1000, 1050)
cagr(1000, 2000, 5)
}
