% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_par_value_tbl.R
\name{create_par_value_tbl}
\alias{create_par_value_tbl}
\title{Present value of investment at various years of maturity}
\usage{
create_par_value_tbl(face_value, rate, years_to_maturity = 0:10)
}
\arguments{
\item{face_value}{Value of money at time t0.}

\item{rate}{Interest rate.}

\item{years_to_maturity}{Number of periods.}
}
\value{
Tibble with present value for each year of maturity
}
\description{
Present value of investment at various years of maturity
}
\examples{
create_par_value_tbl(1000, 0.05)
create_par_value_tbl(1000, 0.10, 0:20)
}
