% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effective_annual_rate.R
\name{effective_annual_rate}
\alias{effective_annual_rate}
\title{Calculate the effective annual rate (EAR) of interest}
\usage{
effective_annual_rate(rate, m)
}
\arguments{
\item{rate}{Interest rate per period}

\item{m}{Number of periods in a year}
}
\value{
Effective annual rate
}
\description{
Calculate the effective annual rate (EAR) of interest
}
\examples{
effective_annual_rate(.10, 2)
}
