% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_value.R
\name{future_value}
\alias{future_value}
\title{Calculate future value of an investment after one or more periods}
\usage{
future_value(present_value, rate, periods)
}
\arguments{
\item{present_value}{Value of money at time t0.}

\item{rate}{Interest rate for money.}

\item{periods}{Number of periods.}
}
\value{
Future value (FV) of the investment.
}
\description{
Calculate future value of an investment after one or more periods
}
\examples{
future_value(1000, 0.05, 3)
future_value(1000, 0.07, 2)
}
