% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/present_value.R
\name{present_value}
\alias{present_value}
\title{Calculate present value of an investment after one or more periods}
\usage{
present_value(future_value, rate, periods)
}
\arguments{
\item{future_value}{Value of money at maturity.}

\item{rate}{Interest rate for money.}

\item{periods}{Number of periods.}
}
\value{
Present value (PV) of the investment.
}
\description{
Calculate present value of an investment after one or more periods
}
\examples{
present_value(1000, 0.05, 3)
present_value(1000, 0.07, 2)
}
