% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/present_value_annuity.R
\name{present_value_annuity}
\alias{present_value_annuity}
\title{Present value of an annuity is the current value of future payments from an annuity}
\usage{
present_value_annuity(payment_c, rate, periods)
}
\arguments{
\item{payment_c}{Aount of future payments.}

\item{rate}{Interest rate.}

\item{periods}{Number of periods.}
}
\value{
Present value per dollar that is received in the future
}
\description{
Present value of an annuity is the current value of future payments from an annuity
}
\examples{
present_value_annuity(1000, 0.05, 20)
present_value_annuity(5000, 0.10, 7)
}
