% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_to_maturity.R
\name{time_to_maturity}
\alias{time_to_maturity}
\title{Calculate the number of periods until maturity for an investment}
\usage{
time_to_maturity(present_value, future_value, rate)
}
\arguments{
\item{present_value}{Value of money at time t0.}

\item{future_value}{Value at maturity.}

\item{rate}{Interest rate.}
}
\value{
Number of periods until maturity
}
\description{
Calculate the number of periods until maturity for an investment
}
\examples{
time_to_maturity(1000, 1050, 0.05)
time_to_maturity(1000, 2000, 0.07)
}
